+++ src/common.ts
@@ -8,6 +8,7 @@ export const PLUGINS_ENABLER_FILENAME = 'SFPluginsTxtEnabler';
 export const NS = 'game-starfield';
 export const GAME_ID = 'starfield';
 export const XBOX_ID = 'BethesdaSoftworks.ProjectGold';
+export const XBOX_APP_X_MANIFEST = 'appxmanifest.xml';
 export const STEAMAPP_ID = '1716740';
 
 export const MOD_TYPE_ASI_MOD = 'starfield-asi-mod';
@@ -16,8 +17,13 @@ export const JUNCTION_NOTIFICATION_ID = 'starfield-junction-notif';
 export const MISSING_PLUGINS_NOTIFICATION_ID = 'starfield-missing-plugins';
 export const MY_GAMES_DATA_WARNING = 'starfield-my-games-data-warning';
 
+// Below constraint is used for the GAME version. Not the extension.
+export const PLUGIN_ENABLER_CONSTRAINT = '<1.12.0';
+
 // This is the order we expect the native plugins to be arranged.
 export const NATIVE_PLUGINS = ['starfield.esm', 'blueprintships-starfield.esm', 'oldmars.esm', 'constellation.esm'];
+export const NATIVE_MID_PLUGINS = ['sfbgs003.esm', 'sfbgs006.esm', 'sfbgs007.esm', 'sfbgs008.esm'];
+export const ALL_NATIVE_PLUGINS = [].concat(NATIVE_PLUGINS, NATIVE_MID_PLUGINS);
 
 export const DATA_SUBFOLDERS = [
   'Meshes',