+++ src/views/InfoPanel.tsx
@@ -17,7 +17,7 @@ interface IConnectedProps {
   discovery: types.IDiscoveryResult;
 }
 
-export default function InfoPanel(props: IBaseProps) {
+export function InfoPanel(props: IBaseProps) {
   const { onInstallPluginsEnabler } = props;
   const { api } = React.useContext(MainContext);
   const [isXbox, setIsXbox] = React.useState(false);
@@ -36,7 +36,7 @@ export default function InfoPanel(props: IBaseProps) {
     if (!isXbox && discovery?.store === 'xbox') {
       setIsXbox(true);
     }
-  }, [pluginEnabler, loadOrder, isXbox, setIsXbox, discovery]);
+  }, [isXbox, pluginEnabler, discovery]);
   const renderXboxWarningEnabledSystem = () => {
     return isXbox ? (
       <Alert bsStyle='warning'>